pyproject.toml
[project]
name = "amelia-ctw-ai-hackathon-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.101.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.1",
]
